+++ golang/go.mod
@@ -6,6 +6,7 @@ require github.com/spf13/cobra v1.6.1
 
 require (
 	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/montanaflynn/stats v0.7.0
 	github.com/spf13/pflag v1.0.5 // indirect
 	gonum.org/v1/gonum v0.12.0
 )
